Refactored the BlockMatcher class. There are now three functions for finding a best match - finding the pixel-accurate best match, refining this to sub-pixel accuracy, and also finding a best match from a list at sub-pixel level. These matching functions contain shortcuts to speed up matching as follows.
FindBestMatchPel:
- The predictors are each searched first, and if their SAD values are small
then the search ends.
- bailout functions are used for calculating matching costs for the remaining
candidates

RefineMatchSubpel:
- The predictor from neighbouring (previously refined) vectors is tested
first. If there's a small SAD, then the search ends
- Pixel-accurate vectors are refined to half-pel vectors by searching the
4 closest half-pel positions (shown as X) first:
 X
XOX
 X

If one of these is better than the pixel-accurate value, then the two nearest
values in the immediate neighbourhood of the pixel-accurate value are also
searched, eg the Y values below:

YXY
XOX
 X

This two-stage procedure is followed again in refining to quarter- and
eighth-pel.
- At each stage of refinement, if we're 10% worse than the value derived
for the predictor, we end the search.

FindBestMatchSubpel:
- bailout functions are used