Changing binarisation for arithmetic coding, and associated contexts. The specification will be modified accordingly.
Binarisation for the magnitude values of wavelet coefficients, motion
vector prediction residues and DC values has been changed from unary to
interleaved exp-Golomb so that the number of symbols to be decoded in order
to reconstruct a coefficient is reduced. Exp-Golomb binarisation takes a
number N>=0 and codes it via the binary representation of N+1, 1bbbbbb. If
there are K bits following the leading 1, the representation is K zeroes
followed by the binary representation:

00...01bbbbb
Interleaved exp-Golomb is the same, except that the K lsbs are interleaved
with the zeroes:
0b0b   0b1
so that a single decoding loop can be used. The zeroes here act as "follow
bits" indicating that another bit is to be sent, with 1 as the terminator.

Contexts for coding these symbols are selected for the follow bits and other
bits ("information bits") separately. Compression performance is hardly
affected, nor is speed performance in software, but hardware performance
is greatly facilitated.