Major restructure of Input/Output classes in Dirac internals. New directory, libdirac_byteio contains classes to input from/output to a Dirac bistream.
Dirac external API (both encoder and decoder) has changed. So the patches
for MPlayer, FFmpeg, and the Dirac Directshow filter will no longer work with
the CVS source. The external API is likely to change over the next few weeks.
Updated versions of the patches and Direcshow filter will be available with the
next release of Dirac.

NOTE: the bytestream has changed to largely conform to the latest Dirac spec.
More work needs to be done to conform fully to the Dirac definition document. So
expect the CVS source and resulting bitstream to change frequently over the next
few weeks.

This patch has been fixed to build under the git repo by davidf 20080702

SKIPPED:
	doc/documentation/code/api/dirac_api.doxygen
	extras/dirac_dshow/src/diracsplitter/DiracSplitter.cpp
	extras/dirac_dshow/src/diracsplitter/DiracSplitter.vcproj
	extras/dirac_dshow/src/diracsplitter/DiracSplitterFile.cpp
	extras/dirac_dshow/src/diracsplitter/stdafx.h