Major refactor of rate control algorithm. The original algorithm aimed for constant bit rate over a single GOP. This tended to be unstable when content changed and could cause quality to degrade significantly at the end of GOPs in order to attempt to correct for high bit rate at the beginning.
The algorithm has been revised so that a target buffer occupancy is aimed
for, with the target bit rate set as equal to the mean bit rate plus an
adjustment factor to steer back to target. Bit rate is computed for a subgroup
(PBB or IBB) in a very similar way, but on a rolling basis rather than a target
for a single GOP.

For the moment, a buffer size of 4xbit rate has been selected. Fixed buffer
sizes may be implemented when levels are specified.

Buffer overflows (too little bit rate - this is a _decoder_ buffer model)
or underflows (too much bit rate) are indicated by warnings in verbose mode.
The buffer occupancy is then limited, so that these events are not corrected
for at present.