Modified rate control. The rate control no longer crashes when it inserts intra pictures: the picture bits in this case are allocated to the _scheduled_ picture type. A tentative buffer model has been re-introduced whereby the GOP allocation is adapted slowly to reduce overshoots, and quickly to reduce undershoots. The buffer target occupancy is set to 90%, and once every GOP this is corrected for bit rate overshoots, and once every subgroup (i.e. much faster) it is corrected, by a larger amount, for bit rate undershoots. In addition, if the bit rate for a particular frame is <50% of the target bit rate, an emergency reallocation is applied.
The frame compressor and sequence compressor classes have been re-arranged
so that the frame compressor manages quantisation selection. This will allow
the rate controller to be modified with look-ahead, changing QF values
ahead of time based on estimated bit rate for residuals, as well as on
the results of motion estimation. At the moment it's still pure feedback
however.

Performance on difficult pictures has also been improved by increasing
the quality of inserted I frames, and increasing the premium given
to pictures with a large number of intra blocks.

Quantiser selection has also been tidied a little.