Actually fix the issue addressed in (082d7a2) (ignored result from fwrite)
Ubuntu forcibly enables FORTIFY_SOURCE, which warns on unused return
values on fwrite.  This may either be fixed by:
 #undef _FORTIFY_SOURCE
 #define _FORTIFY_SOURCE 0

Or, by causing the application to exit on a short write.

NB, it could also be solved by doing a series of fwrites(), and checking
ferror() at the end of them.  FORTIFY_SOURCE would still fail in this
case.